Add passport strateies file
